requirements.txt
beautifulsoup4
pyyaml
progress